fix: resolve three low-severity audit findings
- Finding 14: cwd_slug now appends a hash suffix to prevent collisions
  between distinct paths that map to the same slug
- Finding 22: editor row counts clamp in usize before casting to u16,
  preventing wrap-to-zero for very long content

+++ crates/hrdr-editor/src/lib.rs
@@ -76,7 +76,7 @@ pub trait EditorEngine {
     /// Default counts logical lines (suits no-wrap engines like vim).
     fn desired_rows(&self, _width: u16, max: u16) -> u16 {
         let lines = self.content().split('\n').count().max(1);
-        (lines as u16).clamp(1, max)
+        lines.clamp(1, max as usize) as u16
     }
     /// Insert pasted (bracketed-paste) text at the cursor. Default feeds each
     /// character as a key event (works for any insert-mode engine); engines may
